==> s08_update_variant_annotations/s01_clean_variant_annotations.Rmd <==
---
title: "Ampliseq: clean variant annotations"
author: "Alexey Larionov"
date: "28 Feb 2021"
output:
  html_document:
    toc: true
    number_sections: true
    toc_float: true
editor_options:
  chunk_output_type: console
---

# Summary  

Remove empty and uniform variants annotations: drop the number of variant annotations from 110 to 101.  

Rename variants annotations dataframe: fixed.mx -> vars.mx

# Start section

<style>
pre {
  overflow-x: auto;
}
pre code {
  word-wrap: normal;
  white-space: pre;
}
</style>

```{r echo=F}
options(width = 999)
```

```{r}

# Time stamp
Sys.time()

# Clenan-up
rm(list=ls())
graphics.off()

# Memory
gc()

# Options
options(stringsAsFactors = F,
        warnPartialMatchArgs = T, 
        warnPartialMatchAttr = T, 
        warnPartialMatchDollar = T)

# Files and folders
base_folder="/Users/alexey/Documents/wecare/final_analysis_2021/reanalysis_wo_danish_2021/ampliseq/s01_ampliseq_wecare_only"
data_folder <- file.path(base_folder,"data","s08_update_variant_annotations")
dir.create(data_folder)
scripts_folder <- file.path(base_folder,"scripts","s08_update_variant_annotations")
setwd(scripts_folder)

# Libraries
library(dplyr)

```

# Read data

```{r}

load(file.path(base_folder,"data","s07_qc_filters","r03_filter_cases.RData"))

# Update folders
base_folder="/Users/alexey/Documents/wecare/final_analysis_2021/reanalysis_wo_danish_2021/ampliseq/s01_ampliseq_wecare_only"
data_folder <- file.path(base_folder,"data","s08_update_variant_annotations")
scripts_folder <- file.path(base_folder,"scripts","s08_update_variant_annotations")

```

# Check data

```{r}

ls()

# Dimentions
dim(gt.mx)
dim(fixed.df)

# Consistency
sum(rownames(gt.mx) != rownames(fixed.df))

```

# Explore and rename variants DataFrame

```{r}

# Explore
colnames(fixed.df)
str(fixed.df, list.len=ncol(fixed.df))

# Rename
vars.df <- fixed.df
rm(fixed.df)

```

# Remove empty fields from variants DataFrame

```{r}

# Function to detect empty vectors
all_na.udf <- function(x){all(is.na(x))}

#Detect empty annotations
empty_fields <- apply(vars.df, 2, all_na.udf)
sum(empty_fields)
colnames(vars.df)[empty_fields]

# Remove empty annotations
# VQSLOD and culprit are empty because allele-specific GATK functions were used
vars.df <- vars.df[,!empty_fields]
dim(vars.df)

# Clean-up
rm(all_na.udf, empty_fields)

```

# Check for uniform fields in variants DataFrame

```{r}

# Function to detect uniform vector of an arbitrary datatype
# (assuming no empty and all-NA vectors)
uniform_vector.udf <- function(x)(length(table(x))==1)

# Dtect niform fields 
uniform_fields <- apply(vars.df, 2, uniform_vector.udf)
sum(uniform_fields)
colnames(vars.df)[uniform_fields]
vars.df[1:5,uniform_fields]

# Explore content of uniform fields 
summary(as.factor(vars.df$FILTER))
summary(as.factor(vars.df$AS_FilterStatus))
summary(as.factor(vars.df$exac_non_TCGA.Hemi_AMR))
summary(as.factor(vars.df$exac_non_TCGA.Hemi_EAS))
summary(as.factor(vars.df$exac_non_TCGA.Hemi_OTH))

# Why is this field non-uniform?
summary(as.factor(vars.df$SYMBOL_SOURCE))

# Remove unifirm fields
vars.df <- vars.df %>% 
  select(-FILTER, -AS_FilterStatus, 
         -exac_non_TCGA.Hemi_AMR, -exac_non_TCGA.Hemi_EAS, -exac_non_TCGA.Hemi_OTH)

# Check result
dim(vars.df)
vars.df[1:5,1:5] # Rownames are preserved

# Clean up
rm(uniform_vector.udf, uniform_fields)

```

# Check data

```{r}

ls()

# Dimentions
dim(gt.mx)
dim(vars.df)

# Consistency
sum(rownames(gt.mx) != rownames(vars.df))

```

# Save result

```{r}

save.image(file.path(data_folder, "s01_clean_variant_annotations.RData"))

```

# final_section

```{r final_section}

ls()
sessionInfo()
Sys.time()
gc()

```
